+++ src/main/java/com/github/lmh01/mgt2mt/util/helper/TextAreaHelper.java
@@ -1,6 +1,7 @@
 package com.github.lmh01.mgt2mt.util.helper;
 
 import com.github.lmh01.mgt2mt.util.LogFile;
+import com.github.lmh01.mgt2mt.util.Settings;
 import com.github.lmh01.mgt2mt.windows.WindowMain;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -39,6 +40,9 @@ public class TextAreaHelper {
     }
 
     private static void writeText(String text) {
+        if (Settings.writeTextAreaOutputToConsole) {
+            LOGGER.info(text.trim());
+        }
         WindowMain.TEXT_AREA.append(text);
         WindowMain.TEXT_AREA.append(System.getProperty("line.separator"));
         LogFile.write(text);